Use list.h API in network_server.
server->clients is a `struct list_head` managed by the list API.
A client is added in server_callback_accept() with a malloc() and a list_ad().
server_callback_disconnect() removes the client list_del(), then free the
pointer to the `struct peer_client`.

However:

==19387== Invalid write of size 8
==19387==    at 0x400F69: server_callback_accept (list.h:43)
==19387==    by 0x4E29470: ev_invoke_pending (ev.c:1971)
==19387==    by 0x4E2E0D0: ev_loop (ev.c:2333)
==19387==    by 0x400EA1: server_listen (network_server.c:84)
==19387==    by 0x400CF0: main (main.c:34)
==19387==  Address 0x560a650 is 64 bytes inside a block of size 1,104 free'd
==19387==    at 0x4C21DBC: free (vg_replace_malloc.c:325)
==19387==    by 0x401024: server_callback_read (network_server.c:106)
==19387==    by 0x4E29470: ev_invoke_pending (ev.c:1971)
==19387==    by 0x4E2E0D0: ev_loop (ev.c:2333)
==19387==    by 0x400EA1: server_listen (network_server.c:84)
==19387==    by 0x400CF0: main (main.c:34)

To trigger this bug, set max_clients = 2, start ./example. Connect with to clients:
 $ nc localhost 12345

Then, disconnect both clients, and reconnect with one.

Signed-off-by: Greg Leclercq <ggl@0x80.net>