Add argument chunk_size to simple_buffer_new().
It aims to limit the number of expansize remap() (called by realloc).
There is no constraint on the value of chunk_size but we advise a page size
multiple (you may use getpagesize() to get its value). Depending on the
protocol and data exchanged you might adapt the buffer chunk size.